Delete and Update services created
