+++ Models/Listing.cs
@@ -13,4 +13,5 @@ public class Listing
     public string? Date { get; set; }
     public string? ImageFileName { get; set; }
 
+
 }
