+++ Models/User.cs
@@ -15,5 +15,5 @@ public record class User
     public string PhoneNumber { get; set; } = String.Empty;
     public List<Listing>? Listings { get; set; }
     public bool IsAdmin { get; set; }
-
+    public List<int> Favorites { get; set; } = new List<int>(); 
 }
