+++ Services/IGetListingsService.cs
@@ -5,8 +5,7 @@
         public Task<List<Listing>> GetListings();
         public Task<Listing> GetListing(int id);
         public Task<Listing> GetListing(string Category);
-        public Task DeleteListingById(int id);
-        public Task UpdateListing(int id, SaveListing updatedListing);
+ 
 
 
     }
