+++ Services/IUpdateListingService.cs
@@ -1,0 +1,8 @@
+﻿namespace sahibinden_project.Services
+{
+    public interface IUpdateListingService
+    {
+        public Task UpdateListing(int id, SaveListing updatedListing);
+    }
+
+}
