+++ Services/ListService.cs
@@ -37,29 +37,5 @@ public class ListService : IListService
         return user;
     }
 
-    public async Task Clean()
-    {
-        _db.Users.RemoveRange(_db.Users);
-        await _db.SaveChangesAsync();
-    }
-
-    public async Task DeleteUserById(int id)
-    {
-        var user = await _db.Users.FindAsync(id);
-        if (user != null)
-        {
-            _db.Users.Remove(user);
-            await _db.SaveChangesAsync();
-        }
-    }
-
-    public async Task DeleteUserByUsername(string username)
-    {
-        var user = await _db.Users.SingleOrDefaultAsync(u => u.Username == username);
-        if (user != null)
-        {
-            _db.Users.Remove(user);
-            await _db.SaveChangesAsync();
-        }
-    }
+  
 }
